Make the style editor remove potentially conflicting `vertical-align` styles at applying `<sup>` or `<sub>` style
`execCommand("superscript")` and `execCommand("subscript")` are handled
specially.

`<sup>` and `<sub>` are exclusive. Therefore, existing element is removed first.

However, they are not mapped to CSS due to bug 394304 comment 2 even though the
behavior is different from the other browsers.  Therefore, Gecko does not remove
`vertical-align` styles from selected ranges. On the other hand, the other
browsers do it. This makes sense and matches with the handling in HTML mode.

Therefore, we should remove `vertical-align` style around the selection ranges.

Note that this updates existing WPTs.

The first block of each command checks how the command work with similar
`<span>` having `vertical-align` style. The new expectation matches with the
other browsers (i.e., currently they fail in the other browsers).

The second block of each command checks how the command work with similar
situation but there is a sibling which has the applying style (i.e., `<sup>`
or `<sub>`).  These tests keep failing in the other browsers since they want
to update `vertical-align` style at the selected range in the `<span>` having
`vertical-align` style.

Anyway, this fix is for avoiding new WPT failures of fixing bug 1792386, and
these commands must not be used so widely especially in the CSS mode.
Therefore, I believe that this change is not so risky.

Depends on D163183

Differential Revision: https://phabricator.services.mozilla.com/D163184

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1802418
gecko-commit: 87837958743a87f9c6a99e91b31d3a3b49fc929e
gecko-reviewers: m_kato